Move App\Tests\App\Api to PSR-4 compliant namespace
